add aws tags functions
